+++ src/ProSuite.AGP.WorkList/Domain/Persistence/Json/JsonWorkItemState.cs
@@ -4,7 +4,7 @@ namespace ProSuite.AGP.WorkList.Domain.Persistence.Json
 {
 	public class JsonWorkItemState : IWorkItemState
 	{
-		public int OID { get; set; }
+		public long OID { get; set; }
 		public bool Visited { get; set; }
 		public WorkItemStatus Status { get; set; }
 	}
